+++ src/app/app.module.ts
@@ -24,13 +24,13 @@ import { GithubServiceFactory } from './core/services/factories/factory.github.s
 import { RepoItemServiceFactory } from './core/services/factories/factory.issue.service';
 import { GithubService } from './core/services/github.service';
 import { GithubEventService } from './core/services/githubevent.service';
-import { RepoItemService } from './core/services/issue.service';
+import { RepoItemService } from './core/services/repo-item.service';
 import { LabelService } from './core/services/label.service';
 import { LoggingService } from './core/services/logging.service';
 import { RepoSessionStorageService } from './core/services/repo-session-storage.service';
 import { UserService } from './core/services/user.service';
 import { ViewService } from './core/services/view.service';
-import { IssuesViewerModule } from './issues-viewer/issues-viewer.module';
+import { RepoItemsViewerModule } from './repo-items-viewer/repo-items-viewer.module';
 import { LabelDefinitionPopupComponent } from './shared/label-definition-popup/label-definition-popup.component';
 import { HeaderComponent } from './shared/layout';
 import { RepoChangeFormComponent } from './shared/repo-change-form/repo-change-form.component';
@@ -42,7 +42,7 @@ import { SharedModule } from './shared/shared.module';
     BrowserModule,
     BrowserAnimationsModule,
     AuthModule,
-    IssuesViewerModule,
+    RepoItemsViewerModule,
     ActivityDashboardModule,
     SharedModule,
     HttpClientModule,
