+++ src/app/core/services/factories/factory.auth.service.ts
@@ -6,7 +6,7 @@ import { AuthService } from '../auth.service';
 import { ErrorHandlingService } from '../error-handling.service';
 import { GithubService } from '../github.service';
 import { GithubEventService } from '../githubevent.service';
-import { RepoItemService } from '../issue.service';
+import { RepoItemService } from '../repo-item.service';
 import { LabelService } from '../label.service';
 import { LoggingService } from '../logging.service';
 // import { MockAuthService } from '../mocks/mock.auth.service';
