+++ src/app/repo-items-viewer/card-view/card-view.component.css
