+++ src/app/repo-items-viewer/repo-items-viewer-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from '../core/guards/auth.guard';
-import { IssuesViewerComponent } from './issues-viewer.component';
+import { RepoItemsViewerComponent } from './repo-items-viewer.component';
 
-const routes: Routes = [{ path: 'repoItemsViewer', component: IssuesViewerComponent, canActivate: [AuthGuard] }];
+const routes: Routes = [{ path: 'repoItemsViewer', component: RepoItemsViewerComponent, canActivate: [AuthGuard] }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class IssuesViewerRoutingModule {}
+export class RepoItemsViewerRoutingModule {}
